Fix Loader

+++ CTupeLoader.sh
@@ -6,40 +6,44 @@ if [ -d "CTupeData" ]; then
 fi
 
 wget -P "CTupeData/" https://github.com/nvcuong1312/YtMuos/archive/refs/heads/master.zip
-unzip -o "CTupeData/master.zip" -d "CTupeData/UnzipData/"
+if unzip -o "CTupeData/master.zip" -d "CTupeData/UnzipData/"; then
+	if [ -e "mnt/mmc/MUOS/application/CTupe.sh" ]; then
+		rm -r "mnt/mmc/MUOS/application/CTupe.sh"
+	fi
 
-if [ -e "mnt/mmc/MUOS/application/CTupe.sh" ]; then
-    rm -r "mnt/mmc/MUOS/application/CTupe.sh"
+	cp "CTupeData/UnzipData/YtMuos-master/CTupe.sh" "mnt/mmc/MUOS/application/CTupe.sh"
+
+	if [ -e "usr/bin/youtube-dl" ]; then
+		rm -r "usr/bin/youtube-dl"
+	fi
+
+	cp "CTupeData/UnzipData/YtMuos-master/.ctupe/bin/yt-dlp" "usr/bin/yt-dlp"
+	chmod a+rx /usr/bin/yt-dlp
+	ln -fs /usr/bin/yt-dlp /usr/bin/youtube-dl
+
+	# if [ -e "opt/muos/script/launch/ext-mpv-ctupe.sh" ]; then
+		# rm -r "opt/muos/script/launch/ext-mpv-ctupe.sh"
+	# fi
+
+	# cp "CTupeData/UnzipData/YtMuos-master/.ctupe/data/ext-mpv-ctupe.sh" "opt/muos/script/launch/ext-mpv-ctupe.sh"
+
+	APIKEY="YOUR_API_KEY_HERE"
+	if [ -e "mnt/mmc/MUOS/application/.ctupe/data/API" ]; then
+		APIKEY=$(cat mnt/mmc/MUOS/application/.ctupe/data/API)
+	fi
+
+	if [ -d "mnt/mmc/MUOS/application/.ctupe" ]; then
+	  rm -r "mnt/mmc/MUOS/application/.ctupe"
+	fi
+
+	mv "CTupeData/UnzipData/YtMuos-master/.ctupe" "mnt/mmc/MUOS/application/"
+
+	echo "$APIKEY" > "mnt/mmc/MUOS/application/.ctupe/data/API"
+	
+	echo "Done!"
+else
+	echo "Error!"
 fi
-
-cp "CTupeData/UnzipData/YtMuos-master/CTupe.sh" "mnt/mmc/MUOS/application/CTupe.sh"
-
-if [ -e "usr/bin/youtube-dl" ]; then
-    rm -r "usr/bin/youtube-dl"
-fi
-
-cp "CTupeData/UnzipData/YtMuos-master/.ctupe/bin/yt-dlp" "usr/bin/yt-dlp"
-chmod a+rx /usr/bin/yt-dlp
-ln -fs /usr/bin/yt-dlp /usr/bin/youtube-dl
-
-# if [ -e "opt/muos/script/launch/ext-mpv-ctupe.sh" ]; then
-    # rm -r "opt/muos/script/launch/ext-mpv-ctupe.sh"
-# fi
-
-# cp "CTupeData/UnzipData/YtMuos-master/.ctupe/data/ext-mpv-ctupe.sh" "opt/muos/script/launch/ext-mpv-ctupe.sh"
-
-APIKEY="YOUR_API_KEY_HERE"
-if [ -e "mnt/mmc/MUOS/application/.ctupe/data/API" ]; then
-    APIKEY=$(cat mnt/mmc/MUOS/application/.ctupe/data/API)
-fi
-
-if [ -d "mnt/mmc/MUOS/application/.ctupe" ]; then
-  rm -r "mnt/mmc/MUOS/application/.ctupe"
-fi
-
-mv "CTupeData/UnzipData/YtMuos-master/.ctupe" "mnt/mmc/MUOS/application/"
-
-echo "$APIKEY" > "mnt/mmc/MUOS/application/.ctupe/data/API"
 
 echo "-----------------------------------"
 echo "|Author     : CuongNV             |"
@@ -47,3 +51,4 @@ echo "|Complete!                        |"
 echo "|Thanks!                          |"
 echo "-----------------------------------"
 sleep 3
+
